AG - Add autorization to core methods

+++ api_gateway/app/post_ms/comments/comments_resolvers.py
@@ -2,7 +2,7 @@ import requests
 from app.post_ms.comments.definitions.comments import CommentClass, paginatedComments
 from strawberry.exceptions import GraphQLError
 from app.const import POST_MS_URL
-
+from app.utils.verifyuser import verifyUser
 def get_comment_by_post(PostId:str,page:int) -> paginatedComments:
     try:
         response = requests.get(f"{POST_MS_URL}/comments/?page={page}",headers={"PostId":f"{PostId}"})
@@ -27,7 +27,13 @@ def get_comment_by_post(PostId:str,page:int) -> paginatedComments:
         raise GraphQLError(f"Error al procesar la respuesta: {error}")
 
 
-def create_comment(UserId:str ,PostId:str, Content:str) -> CommentClass:
+def create_comment(token:str ,PostId:str, Content:str) -> CommentClass:
+    userVerified = verifyUser(token)
+    if userVerified == "UNAUTHORIZED":
+        raise GraphQLError("UNAUTHORIZED")
+    if userVerified == "Fallo al verificar":
+        raise GraphQLError("Fallo al verificar")
+    UserId = userVerified.id
     try:
         response = requests.post(f"{POST_MS_URL}/comments", headers = {"UserId":UserId, "PostId":PostId}, json = {"Content":Content})
         print(response.json())
@@ -46,7 +52,12 @@ def create_comment(UserId:str ,PostId:str, Content:str) -> CommentClass:
 
 
 
-def edit_comment(UserId:str ,PostId:str, Content:str, CommentId:str) -> CommentClass:
+def edit_comment(token:str ,PostId:str, Content:str, CommentId:str) -> CommentClass:
+    userVerified = verifyUser(token)
+    if userVerified == "UNAUTHORIZED":
+        raise GraphQLError("UNAUTHORIZED")
+    if userVerified == "Fallo al verificar":
+        raise GraphQLError("Fallo al verificar")
     try:
         response = requests.patch(f"{POST_MS_URL}/comments", headers = {"CommentId":CommentId, "PostId": PostId}, json = {"Content":Content})
         print(response.json())
@@ -63,7 +74,12 @@ def edit_comment(UserId:str ,PostId:str, Content:str, CommentId:str) -> CommentC
     except KeyError as error:  # Keys error
         raise GraphQLError(f"Error al procesar la respuesta: {error}")
     
-def delete_comment(UserId:str ,PostId:str, CommentId:str) -> str:
+def delete_comment(token:str ,PostId:str, CommentId:str) -> str:
+    userVerified = verifyUser(token)
+    if userVerified == "UNAUTHORIZED":
+        raise GraphQLError("UNAUTHORIZED")
+    if userVerified == "Fallo al verificar":
+        raise GraphQLError("Fallo al verificar")
     try:
         response = requests.delete(f"{POST_MS_URL}/comments", headers = {"PostId":PostId,"CommentId":CommentId})
         if response.status_code != 200:
